# api/openapi.yaml
openapi: 3.1.0
...
info:
  title: ProxyLoom P0 API
  version: 1.0.0
  description: >-
    P0 contract for a single configured workspace and trusted administrators.
    Five administrator authentication operations are implemented. Other operations
    remain contract-only; declarations do not assert kernel capability or verification.
    Management uses the session Cookie; internal operations exist exclusively on
    a separate mTLS listener; subscription retrieval uses its mandatory path token
    independently of either credential. Scope and resource IDs are server assigned.
    JSON decoding rejects unknown fields, duplicate keys, trailing documents,
    invalid Unicode, and undocumented nulls. Errors never echo input values.
    All references in this document are local; schema validation must not fetch URLs.
servers:
  - url: /
    description: Management and public subscription listener; internal paths use their own listener.
security:
  - SessionCookie: []
...
components:
  securitySchemes:
    SessionCookie:
      type: "apiKey"
      in: "cookie"
      name: "proxyloom_session"
      description: "Opaque server-side administrator session; Secure, HttpOnly, SameSite=Lax. State-changing management operations additionally enforce Origin and X-CSRF-Token. Setup/login require Origin and JSON but no session CSRF token. Secure may be omitted only for explicit development HTTP loopback PUBLIC_URL. Never accepted as subscription or Runner authorization."
    RunnerMTLS:
      type: "mutualTLS"
      description: "Registered client certificate verified only on the dedicated internal listener. Certificate identity and server registration determine eligible builds, task types and slots. No management Cookie or subscription token is accepted."
  headers:
    RequestID:
      description: "Server-generated safe request correlation ID."
      schema: { $ref: "#/components/schemas/RequestID" }
    ETag:
      description: "Strong resource revision tag \"r<N>\". Exactly one tag, no wildcard, weak tag, list, sign, whitespace or leading zero. N fits positive int64."
      schema:
        type: "string"
        pattern: "^\"r(?:[1-9][0-9]{0,17}|[1-8][0-9]{18}|9[0-1][0-9]{17}|92[0-1][0-9]{16}|922[0-2][0-9]{15}|9223[0-2][0-9]{14}|92233[0-6][0-9]{13}|922337[0-1][0-9]{12}|92233720[0-2][0-9]{10}|922337203[0-5][0-9]{9}|9223372036[0-7][0-9]{8}|92233720368[0-4][0-9]{7}|922337203685[0-3][0-9]{6}|9223372036854[0-6][0-9]{5}|92233720368547[0-6][0-9]{4}|922337203685477[0-4][0-9]{3}|9223372036854775[0-7][0-9]{2}|922337203685477580[0-6]|9223372036854775807)\"$"
    NoStore:
      schema:
        type: "string"
        const: "private, no-store"
    RetryAfter:
      description: "Bounded retry delay in seconds."
      schema:
        type: "integer"
        minimum: 1
        maximum: 86400
  parameters:
...
    ID:
      name: "id"
      in: "path"
      required: true
      schema: { $ref: "#/components/schemas/UUID" }
    IfMatch:
      name: "If-Match"
      in: "header"
      required: true
      description: "Exactly one strong tag \"r<N>\" for the resource being changed. Missing => 428; malformed, weak, wildcard, multi-tag or overflow => 400; valid stale revision => 412. Business/generation/lease/idempotency conflicts => 409. Creates and authentication operations have no resource precondition."
      schema:
        type: "string"
        pattern: "^\"r(?:[1-9][0-9]{0,17}|[1-8][0-9]{18}|9[0-1][0-9]{17}|92[0-1][0-9]{16}|922[0-2][0-9]{15}|9223[0-2][0-9]{14}|92233[0-6][0-9]{13}|922337[0-1][0-9]{12}|92233720[0-2][0-9]{10}|922337203[0-5][0-9]{9}|9223372036[0-7][0-9]{8}|92233720368[0-4][0-9]{7}|922337203685[0-3][0-9]{6}|9223372036854[0-6][0-9]{5}|92233720368547[0-6][0-9]{4}|922337203685477[0-4][0-9]{3}|9223372036854775[0-7][0-9]{2}|922337203685477580[0-6]|9223372036854775807)\"$"
    IdempotencyKey:
      name: "Idempotency-Key"
      in: "header"
      required: false
      description: "Supported on asynchronous creation, publication, token issuance and import commit. One key per authenticated actor + fixed route; request HMAC also binds scope and canonical body. Same key/body waits then replays permitted metadata; changed body is 409. Mutations and idempotency registration commit in the same transaction. No arbitrary response bytes or one-time secrets are replayed."
      schema:
        type: "string"
        pattern: "^[A-Za-z0-9_-]{1,128}$"
    Limit:
      name: "limit"
      in: "query"
      schema:
        type: "integer"
        minimum: 1
        maximum: 200
        default: 50
      description: "Bounded page size. Stable ascending (created_at, id) order unless a collection documents its immutable equivalent."
    Cursor:
      name: "cursor"
      in: "query"
      schema:
        type: "string"
        minLength: 1
        maxLength: 2048
      description: "Opaque authenticated continuation tied to scope, collection, order and normalized filter digest. A changed filter or invalid cursor is 400; authorization is checked separately on every request."
...
  responses:
    BadRequest:
      description: "Malformed JSON, duplicate/unknown field, unsafe shape, invalid path/query/header, invalid Unicode or cursor. No submitted key or value is echoed."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: "Required management session or registered internal identity is missing, invalid or expired; surfaces are not interchangeable."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: "Authenticated actor lacks permission, recent authentication, a valid current-actor confirmation or required same-origin protection."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: "Resource does not exist in the authorized scope, or existence must not be disclosed."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Conflict:
      description: "Business state, generation, idempotency or lease conflict, including COMPILE_OBSOLETE, IDEMPOTENCY_CONFLICT and LEASE_LOST. A stale resource revision is 412 instead."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    PreconditionFailed:
      description: "A syntactically valid If-Match names an outdated resource revision; no mutation was applied."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    TooLarge:
      description: "Input, decompressed data, item count or event size exceeds the configured bound."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    UnprocessableEntity:
      description: "Shape is understood but merged domain semantics are invalid, including required secret clear, unsupported capability, invalid references, excluded dependencies and DNS cycles."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    PreconditionRequired:
      description: "The versioned resource operation requires If-Match but it was omitted."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    RateLimited:
      description: "Rate, slot, concurrent execution or reserved budget limit exceeded."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
        Retry-After: { $ref: "#/components/headers/RetryAfter" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unavailable:
      description: "Primary database, execution capability or safe publication is unavailable. Fail closed; never serve cached authorization or substitute a direct/empty configuration."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
        Retry-After: { $ref: "#/components/headers/RetryAfter" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    ImportAccepted:
      description: "Typed contract response. This operation is not registered by the foundation contract implementation."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
        ETag: { $ref: "#/components/headers/ETag" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ImportAcceptedResponse" }
...
    ImportCommitResponse:
      description: "Typed contract response. This operation is not registered by the foundation contract implementation."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
        ETag: { $ref: "#/components/headers/ETag" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ImportCommitResponse" }
    ImportResponse:
      description: "Typed contract response. This operation is not registered by the foundation contract implementation."
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestID" }
        Cache-Control: { $ref: "#/components/headers/NoStore" }
        ETag: { $ref: "#/components/headers/ETag" }
      content:
        "application/json":
          schema: { $ref: "#/components/schemas/ImportResponse" }
...
  schemas:
    UUID:
      type: "string"
      pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-[1-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
    RequestID:
      type: "string"
      pattern: "^[A-Za-z0-9_-]{1,64}$"
    Revision:
      type: "string"
      pattern: "^(?:[1-9][0-9]{0,17}|[1-8][0-9]{18}|9[0-1][0-9]{17}|92[0-1][0-9]{16}|922[0-2][0-9]{15}|9223[0-2][0-9]{14}|92233[0-6][0-9]{13}|922337[0-1][0-9]{12}|92233720[0-2][0-9]{10}|922337203[0-5][0-9]{9}|9223372036[0-7][0-9]{8}|92233720368[0-4][0-9]{7}|922337203685[0-3][0-9]{6}|9223372036854[0-6][0-9]{5}|92233720368547[0-6][0-9]{4}|922337203685477[0-4][0-9]{3}|9223372036854775[0-7][0-9]{2}|922337203685477580[0-6]|9223372036854775807)$"
      description: "Canonical positive decimal int64 string, 1..9223372036854775807. String encoding preserves exact values in browser clients. No signs or leading zeros."
...
    Name:
      type: "string"
      minLength: 1
      maxLength: 256
      pattern: "^[^\\x00-\\x1f\\x7f]+$"
    Tags:
      type: "array"
      minItems: 0
      maxItems: 100
      items: { $ref: "#/components/schemas/Tag" }
      uniqueItems: true
...
    Timestamp:
      type: "string"
      format: "date-time"
    SafeFieldPath:
      type: "string"
      maxLength: 512
      pattern: "^(?:/(?:[a-z][a-z0-9_]*|[0-9]+))*$"
      description: "Safe RFC 6901 pointer using only fixed, validated contract fields and array indices; root is empty. Unknown or duplicate input keys are represented by their safe parent, never echoed. Syntax alone is not authorization to output a supplied path."
    ErrorCode:
      type: "string"
      enum: ["MALFORMED_REQUEST", "UNKNOWN_FIELD", "DUPLICATE_FIELD", "AUTH_REQUIRED", "SESSION_EXPIRED", "PERMISSION_DENIED", "REAUTH_REQUIRED", "RESOURCE_NOT_FOUND", "STATE_CONFLICT", "IDEMPOTENCY_CONFLICT", "LEASE_LOST", "REVISION_MISMATCH", "INPUT_LIMIT_EXCEEDED", "VALIDATION_FAILED", "PRECONDITION_REQUIRED", "RATE_LIMITED", "SERVICE_UNAVAILABLE", "INTERNAL_ERROR", "COMPILE_OBSOLETE", "CAPABILITY_UNSUPPORTED", "CAPABILITY_UNVERIFIED", "DEPENDENCY_EXCLUDED", "DEPENDENCY_CYCLE", "DNS_CYCLE", "EMPTY_GROUP", "PUBLICATION_BLOCKED", "PUBLICATION_CONFIRMATION_REQUIRED", "SUBSCRIPTION_NOT_READY", "SUBSCRIPTION_BLOCKED", "RUNNER_UNAVAILABLE", "BUDGET_EXCEEDED", "INVALID_CONFIG", "CORE_CONFIG_INVALID", "PROXY_CONNECT_FAILED", "AUTH_FAILED", "TARGET_TLS_FAILED", "HTTP_EXPECTATION_FAILED", "CORE_BINARY_MISMATCH", "PORT_BUSY", "RUNNER_RESOURCE_LIMIT", "CANCELED", "JOB_TIMEOUT", "TEST_TARGET_UNAVAILABLE", "INSUFFICIENT_SAMPLE"]
    ErrorDetail:
      type: "object"
      additionalProperties: false
      required: []
      properties:
        field_path: { $ref: "#/components/schemas/SafeFieldPath" }
        resource_id: { $ref: "#/components/schemas/UUID" }
      minProperties: 1
    SafeError:
      type: "object"
      additionalProperties: false
      required: ["code", "message", "details"]
      properties:
        code: { $ref: "#/components/schemas/ErrorCode" }
        message:
          type: "string"
          minLength: 1
          maxLength: 512
          description: "Bounded fixed safe explanation; no input values, SQL, ciphertext, URLs, raw schema messages or core stdout/stderr."
        details:
          type: "array"
          minItems: 0
          maxItems: 64
          items: { $ref: "#/components/schemas/ErrorDetail" }
    ErrorResponse:
      type: "object"
      additionalProperties: false
      required: ["request_id", "error"]
      properties:
        request_id: { $ref: "#/components/schemas/RequestID" }
        error: { $ref: "#/components/schemas/SafeError" }
...
    Diagnostics:
      type: "array"
      minItems: 0
      maxItems: 200
      items: { $ref: "#/components/schemas/Diagnostic" }
    PageInfo:
      type: "object"
      additionalProperties: false
      required: ["limit"]
      properties:
        limit:
          type: "integer"
          minimum: 1
          maximum: 200
          default: 50
        next_cursor:
          type: "string"
          minLength: 1
          maxLength: 2048
          description: "Opaque authenticated cursor; omitted at end. Bound to scope, collection, ordering and normalized filter digest. Cursor is not authorization."
...
    Endpoint:
      type: "object"
      additionalProperties: false
      required: ["host", "port"]
      properties:
        host: { $ref: "#/components/schemas/Host" }
        port:
          type: "integer"
          minimum: 1
          maximum: 65535
    Protocol:
      type: "string"
      enum: ["shadowsocks", "vmess", "vless", "trojan", "socks5", "http"]
...
    NodePatch:
      type: "object"
      additionalProperties: false
      required: []
      properties:
        protocol: { $ref: "#/components/schemas/Protocol" }
        endpoint: { $ref: "#/components/schemas/Endpoint" }
        auth: { $ref: "#/components/schemas/NodeAuthPatch" }
        transport: { $ref: "#/components/schemas/NodeTransport" }
        security: { $ref: "#/components/schemas/NodeSecurityPatch" }
        features: { $ref: "#/components/schemas/NodeFeatures" }
      minProperties: 1
      description: "Only these typed fields are writable. Endpoint/transport/features are whole-object replacements; auth/security use the documented typed merge. No arbitrary merge-patch, scope, revision, epoch, origin or native dependency fields."
...
    NodePatchRequest:
      type: "object"
      additionalProperties: false
      required: []
      properties:
        name: { $ref: "#/components/schemas/Name" }
        tags: { $ref: "#/components/schemas/Tags" }
        enabled:
          type: "boolean"
        node: { $ref: "#/components/schemas/NodePatch" }
      minProperties: 1
...
    ImportFormat:
      type: "string"
      enum: ["auto", "uri_list", "base64_uri_list", "xray_json", "singbox_json", "mihomo_yaml"]
    ImportCreateRequest:
      type: "object"
      additionalProperties: false
      required: ["text", "format"]
      properties:
        text:
          type: "string"
          minLength: 1
          maxLength: 10485760
          x-sensitive: true
          description: "Bounded source text; never logged. Parsing and decoded-size limits also apply."
        format: { $ref: "#/components/schemas/ImportFormat" }
        source_id: { $ref: "#/components/schemas/UUID" }
    ImportFileRequest:
      type: "object"
      additionalProperties: false
      required: ["file", "format"]
      properties:
        file:
          type: "string"
          format: "binary"
          maxLength: 10485760
          description: "Restricted file bytes only; the client filename is never a server filesystem path."
        format: { $ref: "#/components/schemas/ImportFormat" }
        source_id: { $ref: "#/components/schemas/UUID" }
    ImportAccepted:
      type: "object"
      additionalProperties: false
      required: ["batch_id", "job_id", "revision", "state"]
      properties:
        batch_id: { $ref: "#/components/schemas/UUID" }
        job_id: { $ref: "#/components/schemas/UUID" }
        revision: { $ref: "#/components/schemas/Revision" }
        state:
          type: "string"
          enum: ["queued", "parsing"]
    ImportAcceptedResponse:
      type: "object"
      additionalProperties: false
      required: ["request_id", "data"]
      properties:
        request_id: { $ref: "#/components/schemas/RequestID" }
        data: { $ref: "#/components/schemas/ImportAccepted" }
    ImportCandidate:
      type: "object"
      additionalProperties: false
      required: ["candidate_id", "index", "state", "diagnostics"]
      properties:
        candidate_id: { $ref: "#/components/schemas/UUID" }
        index:
          type: "integer"
          minimum: 0
          maximum: 99999
        state:
          type: "string"
          enum: ["new", "matched", "conflict", "invalid"]
        name: { $ref: "#/components/schemas/Name" }
        node: { $ref: "#/components/schemas/NodeRedacted" }
        existing_resource_id: { $ref: "#/components/schemas/UUID" }
        existing_revision: { $ref: "#/components/schemas/Revision" }
        match_method:
          type: "string"
          enum: ["stable_external_key", "manual_binding", "exact_fingerprint", "suggestion"]
        changed_fields:
          type: "array"
          minItems: 0
          maxItems: 64
          items: { $ref: "#/components/schemas/SafeFieldPath" }
        diagnostics: { $ref: "#/components/schemas/Diagnostics" }
    ImportBatch:
      type: "object"
      additionalProperties: false
      required: ["batch_id", "revision", "job_id", "state", "candidate_count", "candidates", "created_at", "expires_at", "diagnostics"]
      properties:
        batch_id: { $ref: "#/components/schemas/UUID" }
        revision: { $ref: "#/components/schemas/Revision" }
        job_id: { $ref: "#/components/schemas/UUID" }
        state:
          type: "string"
          enum: ["queued", "parsing", "ready", "failed", "committed", "expired"]
        source_id: { $ref: "#/components/schemas/UUID" }
        source_revision: { $ref: "#/components/schemas/Revision" }
        binding_revision: { $ref: "#/components/schemas/Revision" }
        expires_at: { $ref: "#/components/schemas/Timestamp" }
        candidate_count:
          type: "integer"
          minimum: 0
          maximum: 100000
        candidates:
          type: "array"
          minItems: 0
          maxItems: 200
          items: { $ref: "#/components/schemas/ImportCandidate" }
        created_at: { $ref: "#/components/schemas/Timestamp" }
        diagnostics: { $ref: "#/components/schemas/Diagnostics" }
    ImportResponse:
      type: "object"
      additionalProperties: false
      required: ["request_id", "data", "page"]
      properties:
        request_id: { $ref: "#/components/schemas/RequestID" }
        data: { $ref: "#/components/schemas/ImportBatch" }
        page: { $ref: "#/components/schemas/PageInfo" }
    ImportDecision:
      type: "object"
      additionalProperties: false
      required: ["candidate_id", "action"]
      properties:
        candidate_id: { $ref: "#/components/schemas/UUID" }
        action:
          type: "string"
          enum: ["create", "update", "skip", "bind"]
        resource_id: { $ref: "#/components/schemas/UUID" }
        expected_revision: { $ref: "#/components/schemas/Revision" }
        override: { $ref: "#/components/schemas/NodePatchRequest" }
      allOf:
        -
          if:
            properties:
              action:
                enum: ["update", "bind"]
          then:
            required: ["resource_id", "expected_revision"]
        -
          if:
            properties:
              action:
                enum: ["create", "skip"]
          then:
            not:
              anyOf:
                -
                  required: ["resource_id"]
                -
                  required: ["expected_revision"]
    ImportCommitRequest:
      type: "object"
      additionalProperties: false
      required: ["decisions"]
      properties:
        decisions:
          type: "array"
          minItems: 1
          maxItems: 5000
          items: { $ref: "#/components/schemas/ImportDecision" }
        source_revision: { $ref: "#/components/schemas/Revision" }
        binding_revision: { $ref: "#/components/schemas/Revision" }
      description: "The preview ETag is supplied in If-Match. Source-backed commits additionally require matching source_revision and binding_revision. Selected writes and binding changes are one transaction; conflicts never silently overwrite."
    ImportCommitItem:
      type: "object"
      additionalProperties: false
      required: ["candidate_id", "status"]
      properties:
        candidate_id: { $ref: "#/components/schemas/UUID" }
        status:
          type: "string"
          enum: ["created", "updated", "skipped", "bound"]
        resource_id: { $ref: "#/components/schemas/UUID" }
        revision: { $ref: "#/components/schemas/Revision" }
    ImportCommit:
      type: "object"
      additionalProperties: false
      required: ["batch_id", "revision", "items"]
      properties:
        batch_id: { $ref: "#/components/schemas/UUID" }
        revision: { $ref: "#/components/schemas/Revision" }
        items:
          type: "array"
          minItems: 1
          maxItems: 5000
          items: { $ref: "#/components/schemas/ImportCommitItem" }
    ImportCommitResponse:
      type: "object"
      additionalProperties: false
      required: ["request_id", "data"]
      properties:
        request_id: { $ref: "#/components/schemas/RequestID" }
        data: { $ref: "#/components/schemas/ImportCommit" }
...
    Tag:
      type: "string"
      minLength: 1
      maxLength: 64
      pattern: "^[^\\x00-\\x1f\\x7f]+$"
paths:
...
  /api/v1/imports:
    post:
      operationId: createImport
      x-implementation-status: implemented
      summary: Create a bounded import preview job on the API Worker
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ImportCreateRequest'}
          multipart/form-data:
            schema: {$ref: '#/components/schemas/ImportFileRequest'}
      responses:
        '202': {$ref: '#/components/responses/ImportAccepted'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
        '413': {$ref: '#/components/responses/TooLarge'}
        '422': {$ref: '#/components/responses/UnprocessableEntity'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /api/v1/imports/{id}:
    parameters: [{$ref: '#/components/parameters/ID'}]
    get:
      operationId: getImport
      x-implementation-status: implemented
      summary: Read redacted import candidates and per-entry diagnostics
      parameters: [{$ref: '#/components/parameters/Limit'}, {$ref: '#/components/parameters/Cursor'}]
      responses:
        '200': {$ref: '#/components/responses/ImportResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /api/v1/imports/{id}/commit:
    parameters: [{$ref: '#/components/parameters/ID'}]
    post:
      operationId: commitImport
      x-implementation-status: implemented
      summary: Commit selected candidates transactionally against the preview revision
      parameters: [{$ref: '#/components/parameters/IfMatch'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ImportCommitRequest'}
      responses:
        '200': {$ref: '#/components/responses/ImportCommitResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '412': {$ref: '#/components/responses/PreconditionFailed'}
        '422': {$ref: '#/components/responses/UnprocessableEntity'}
        '428': {$ref: '#/components/responses/PreconditionRequired'}
        '503': {$ref: '#/components/responses/Unavailable'}
...